Move JwtTokenUser into the same namespace as the related interface
